Move history to own module
